+++ packages/wallet-core/.prettierignore
