+++ packages/wallet-core/src/utils.ts
@@ -18,16 +18,3 @@ export function withResolvers<T = any>(): Resolvers<T> {
 
 	return { promise, reject: reject!, resolve: resolve! };
 }
-
-export function getClientMetadata() {
-	return {
-		version: 'v1',
-		originUrl: window.location.href,
-		userAgent: navigator.userAgent,
-		screenResolution: `${window.screen.width}x${window.screen.height}`,
-		language: navigator.language,
-		platform: navigator.platform,
-		timezone: Intl.DateTimeFormat().resolvedOptions().timeZone,
-		timestamp: Date.now(),
-	};
-}
