+++ packages/wallet-core/src/web-wallet-channel/index.ts
@@ -1,0 +1,5 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+export { WalletPostMessageChannel } from './wallet-post-message-channel';
+export { DappPostMessageChannel } from './dapp-post-message-channel';
