+++ packages/wallet-core/src/web-wallet-channel/utils.ts
@@ -1,0 +1,12 @@
+export function getClientMetadata() {
+	return {
+		version: 'v1',
+		originUrl: window.location.href,
+		userAgent: navigator.userAgent,
+		screenResolution: `${window.screen.width}x${window.screen.height}`,
+		language: navigator.language,
+		platform: navigator.platform,
+		timezone: Intl.DateTimeFormat().resolvedOptions().timeZone,
+		timestamp: Date.now(),
+	};
+}
