+++ packages/wallet-core/tsconfig.esm.json
