+++ packages/wallet-core/tsconfig.json
