+++ packages/wallet-core/typedoc.json
