+++ packages/wallet-core/vitest.config.mts
