--- packages/window-wallet-channel/src/index.ts
@@ -1,5 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-export { HostResponse } from './host-response.js';
-export { WindowRequest as PopupRequest } from './window-request.js';
